게시글 api 설계완료

+++ src/main/java/com/sparta/springhw01/service/PostService.java
@@ -1,12 +1,18 @@
 package com.sparta.springhw01.service;
 
 import com.sparta.springhw01.dto.PostRequestDto;
+import com.sparta.springhw01.dto.PostResponseDto;
 import com.sparta.springhw01.entity.Post;
+import com.sparta.springhw01.entity.User;
+import com.sparta.springhw01.jwt.JwtUtil;
 import com.sparta.springhw01.repository.PostRepository;
+import com.sparta.springhw01.repository.UserRepository;
+import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 
 @Service
@@ -14,12 +20,36 @@ import java.util.List;
 public class PostService {
 
     private final PostRepository postRepository;
+    private final UserRepository userRepository;
+    private final JwtUtil jwtUtil;
 
     @Transactional
-    public Post createPost(PostRequestDto requestDto) {
-        Post post = new Post(requestDto);
-        postRepository.save(post);
-        return post;
+    public PostResponseDto createPost(PostRequestDto requestDto, HttpServletRequest request) {
+        // Request에서 Token 가져오기
+        String token = jwtUtil.resolveToken(request);
+        Claims claims;
+
+        // 토큰이 있는 경우에만 관심상품 추가 가능
+        if (token != null) {
+            if (jwtUtil.validateToken(token)) {
+                // 토큰에서 사용자 정보 가져오기
+                claims = jwtUtil.getUserInfoFromToken(token);
+            } else {
+                throw new IllegalArgumentException("Token Error");
+            }
+
+            // 토큰에서 가져온 사용자 정보를 사용하여 DB 조회
+            User user = userRepository.findByUsername(claims.getSubject()).orElseThrow(
+                    () -> new IllegalArgumentException("사용자가 존재하지 않습니다.")
+            );
+            System.out.println("dfdf 나오나??"+user.getId());
+            // 요청받은 DTO 로 DB에 저장할 객체 만들기
+            Post post = postRepository.saveAndFlush(new Post(requestDto, user.getId()));
+
+            return new PostResponseDto(post);
+        } else {
+            return null;
+        }
     }
 
     @Transactional(readOnly = true)
@@ -28,27 +58,70 @@ public class PostService {
     }
 
     @Transactional
-    public Post update(Long id, PostRequestDto requestDto) {
-        Post post = postRepository.findById(id).orElseThrow(
-                () -> new IllegalArgumentException("존재하지 않는 게시글입니다.")
-        );
+    public Post update(Long id, PostRequestDto requestDto, HttpServletRequest request) {
+        // Request에서 Token 가져오기
+        String token = jwtUtil.resolveToken(request);
+        Claims claims;
 
-        postRepository.findByIdAndPassword(id, requestDto.getPassword()).orElseThrow(
-                () -> new IllegalArgumentException("패스워드가 일치하지 않습니다.")
-        );
+        // 토큰이 있는 경우에만 관심상품 최저가 업데이트 가능
+        if (token != null) {
+            // Token 검증
+            if (jwtUtil.validateToken(token)) {
+                // 토큰에서 사용자 정보 가져오기
+                claims = jwtUtil.getUserInfoFromToken(token);
+            } else {
+                throw new IllegalArgumentException("Token Error");
+            }
 
-        post.update(requestDto);
-        return post;
+            // 토큰에서 가져온 사용자 정보를 사용하여 DB 조회
+            User user = userRepository.findByUsername(claims.getSubject()).orElseThrow(
+                    () -> new IllegalArgumentException("사용자가 존재하지 않습니다.")
+            );
+
+            Post post = postRepository.findByIdAndUserId(id, user.getId()).orElseThrow(
+                    () -> new NullPointerException("해당 게시글은 존재하지 않습니다.")
+            );
+
+            post.update(requestDto);
+
+            return post;
+
+        } else {
+            return null;
+        }
     }
 
     @Transactional
-    public boolean deletePost(Long id, PostRequestDto requestDto) {
-        postRepository.findByIdAndPassword(id, requestDto.getPassword()).orElseThrow(
-                () -> new IllegalArgumentException("패스워드가 일치하지 않습니다.")
-        );
+    public boolean deletePost(Long id, PostRequestDto requestDto, HttpServletRequest request) {
+        // Request에서 Token 가져오기
+        String token = jwtUtil.resolveToken(request);
+        Claims claims;
 
-        postRepository.deleteById(id);
-        return true;
+        // 토큰이 있는 경우에만 관심상품 최저가 업데이트 가능
+        if (token != null) {
+            // Token 검증
+            if (jwtUtil.validateToken(token)) {
+                // 토큰에서 사용자 정보 가져오기
+                claims = jwtUtil.getUserInfoFromToken(token);
+            } else {
+                throw new IllegalArgumentException("Token Error");
+            }
+
+            // 토큰에서 가져온 사용자 정보를 사용하여 DB 조회
+            User user = userRepository.findByUsername(claims.getSubject()).orElseThrow(
+                    () -> new IllegalArgumentException("사용자가 존재하지 않습니다.")
+            );
+
+            Post post = postRepository.findByIdAndUserId(id, user.getId()).orElseThrow(
+                    () -> new NullPointerException("해당 게시글은 존재하지 않습니다.")
+            );
+
+            postRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
+
     }
 
     @Transactional(readOnly = true)
